Extract common filter to base to reduce code
