+++ robinho/classifiers/extremely_biased.py
@@ -11,9 +11,6 @@ from sklearn.linear_model import SGDClassifier
 class ExtremelyBiased(BaseClassifier):
     name = "extremely_biased"
 
-    def filter(self, df):
-        return df['content'].str.len() > 120
-
     def features_labels(self):
         df = self.load_links()
